+++ barghos-math/src/main/java/org/barghos/math/matrix/Mat3.java
@@ -183,7 +183,7 @@ public class Mat3
 	public float determinant()
 	{
 		assert(m != null);
-		return MatUtils.det3x3( this.m[0][0], this.m[0][1], this.m[0][2],
+		return (float) MatUtils.det3x3( this.m[0][0], this.m[0][1], this.m[0][2],
 								this.m[1][0], this.m[1][1], this.m[1][2],
 								this.m[2][0], this.m[2][1], this.m[2][2]);
 	}
